initialize opt if not provided in inference()
